Move test fixtures for stac tests into conftest.py
